fixed bug when sent transactions were showed as Received
removed unused imports

only parsing the first 3 txs

outgoing txs have 3 parts, hence to get the last 3 tx we need to slice up to position 9
